+++ scripts/update-progress.js
@@ -4,12 +4,14 @@ const date_options = { year: 'numeric', month: 'long', day: 'numeric' };
 
 function update_progress()
 {
-    const normal_quest_updated = false;
-    const hard_quest_updated = false;
+    const normal_quest_updated = true;
+    const hard_quest_updated = true;
     const equipment_selection_updated = true;
     const english_translated_names_updated = true;
     const rank_equipment_updated = false;
-    // also don't forget to update crap like README lol.
+    // and README
+    // and Settings Quest Min/Max
+    // and Presets Min/Max
 
     if (!normal_quest_updated || !hard_quest_updated || !equipment_selection_updated || !english_translated_names_updated || !rank_equipment_updated)
     {